extract appiums, regroup
