+++ utils/utils.go
@@ -5,14 +5,19 @@ import (
 	"log"
 	"net"
 	"net/http"
+	"time"
+
+	"github.com/frezot/appioid/settings"
 )
+
+const localhost = "http://127.0.0.1:"
 
 // BuildAppioidBaseURL detects ip and concat with port to URL
 func BuildAppioidBaseURL(port string) string {
 	conn, err := net.Dial("udp", "8.8.8.8:80")
 	if err != nil {
 		log.Println("[DEBUG] ip autodetect failed. localhost will be used")
-		return "http://127.0.0.1:" + port
+		return localhost + port
 	}
 	defer conn.Close()
 	localAddr := conn.LocalAddr().(*net.UDPAddr)
@@ -21,7 +26,7 @@ func BuildAppioidBaseURL(port string) string {
 
 // AppiumServerURL build URL for interact with appium
 func AppiumServerURL(port string) string {
-	return "http://127.0.0.1:" + port + "/wd/hub"
+	return localhost + port + "/wd/hub"
 }
 
 // AppiumStatus return current atatus of Appium server
@@ -33,4 +38,19 @@ func AppiumStatus(port string) string {
 		return string(responseData)
 	}
 	return "ERR"
+}
+
+// AppiumIsReady gently get status and returns as boolean
+func AppiumIsReady(port string) bool {
+
+	singleLatency := 6 //experimentally established value
+
+	for i := 0; i < singleLatency*settings.PoolSize; i++ {
+		if AppiumStatus(port) == "ERR" {
+			time.Sleep(500 * time.Millisecond)
+		} else {
+			return true
+		}
+	}
+	return (AppiumStatus(port) != "ERR")
 }
